pyproject.toml
[project]
name = "gptoss-bullshit-detector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.10.0",
    "fire>=0.7.0",
    "jaxtyping>=0.3.2",
    "kernels>=0.9.0",
    "matplotlib>=3.10.5",
    "nnsight>=0.4.11",
    "openai>=1.100.2",
    "openai-harmony>=0.0.4",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.1",
    "tqdm>=4.67.1",
    "transformers>=4.55.0",
]
